requirements.txt
python-dotenv
python-telegram-bot
langchain
langchain-community
langchain-core
faiss-cpu
sentence_transformers
openai
tools

